refactor: modularize relational-select tool into folder structure
- Break down monolithic relational-select.ts (316 lines) into focused modules
- Create separate files for types, schemas, query-builder, executor, and main tool
- Refactor tests into modular structure with separate files for utilities, schema validation, and tool invocation
- Maintain all existing functionality and test coverage (22 tests passing)
- Improve code organization and maintainability

Source modules:
- types.ts: TypeScript type definitions and interfaces
- schemas.ts: Zod validation schemas
- query-builder.ts: SQL query construction logic
- executor.ts: Query execution and error handling
- index.ts: Main tool definition and exports

Test modules:
- test-utils.ts: Shared test utilities
- schema-validation.test.ts: Schema validation tests
- tool-invocation.test.ts: Integration tests
- index.test.ts: Main test entry point
