+++ packages/tools/src/data/relational/tools/index.ts
@@ -4,5 +4,5 @@
  */
 
 export { relationalQuery } from './relational-query.js';
-export { relationalSelect } from './relational-select.js';
+export { relationalSelect } from './relational-select/index.js';
