+++ packages/tools/tests/data/relational/relational-select/index.test.ts
@@ -1,0 +1,11 @@
+/**
+ * Relational SELECT Tool Tests
+ *
+ * Main test suite that imports all relational-select test modules.
+ * Tests tool invocation, schema validation, WHERE conditions, ORDER BY, and pagination.
+ */
+
+// Import all test suites
+import './schema-validation.test.js';
+import './tool-invocation.test.js';
+
