+++ packages/tools/tests/data/relational/relational-select/test-utils.ts
@@ -1,0 +1,22 @@
+/**
+ * Test utilities for relational SELECT tests
+ * @module tests/relational-select/test-utils
+ */
+
+/**
+ * Check if better-sqlite3 native bindings are available
+ * This is used to conditionally skip tests that require SQLite
+ */
+export const hasSQLiteBindings = (() => {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-require-imports
+    const Database = require('better-sqlite3');
+    // Try to actually instantiate it with an in-memory database
+    const db = new Database(':memory:');
+    db.close();
+    return true;
+  } catch {
+    return false;
+  }
+})();
+
